fix overlap layer and tests
There was a bug with getting a mask from None --- I suppose this means that we must always assume that masks that are not provided are None. Not sure if this is a change from Keras 2.0, or if it was never supposed to work.
There's also an odd case in the test_masked_batched_case function --- for some reason, it thinks that __call__ takes 2 arguments. There is probably a more sinister bug under this, so we should look into it.